requirements.txt
numpy
pandas
opencv-python
networkx
h5py
tqdm
vocab
revtok
Pillow
torch==1.6.0
torchvision==0.7.0
tensorboardX==1.8
ai2thor==2.1.0
